+++ src/searchbar/index.js
@@ -1,1 +1,0 @@
-export SearchBar from './searchbar.component.jsx';