+++ webpack/base.config.js
@@ -33,7 +33,13 @@ function getBaseConfiguration(config) {
         },
         {
           test: /\.svg$/,
+          exclude: path.resolve(__dirname, '..', 'node_modules', 'font-awesome'),
           use: ['babel-loader', 'react-svg-loader'],
+        },
+        {
+          test: /\.svg$/,
+          include: path.resolve(__dirname, '..', 'node_modules', 'font-awesome'),
+          use: ['file-loader?name=[name].[ext]'],
         },
         {
           test: /\.ejs$/,